Refactor: move employee add/edit page into its own component
