+++ onboarding/src/Components/EmployeeEdit/UserManagerProfilePage.js
@@ -3,10 +3,10 @@ import React, { useRef } from "react";
 //import "../static/looper/stylesheets/theme.min.css";
 //import "../static/looper/stylesheets/theme-dark.min.css";
 
-import UserProfileManage from "./UserProfileManage";
-import Navbar from "./Navbar";
-import LeftMenu from "./LeftMenu";
-import PageAddressBar from "./PageAddressBar"
+import EmployeeEditForm from "./EmployeeEditForm";
+import Navbar from "../Navbar";
+import LeftMenu from "../LeftMenu";
+import PageAddressBar from "../PageAddressBar"
 
 function UserManagerProfilePage(props) {
     const packageIdRef = useRef(0);
@@ -52,20 +52,7 @@ function UserManagerProfilePage(props) {
     					<div className="page-section">
     						<div className="card card-fluid">
     							<div className="card-header">Pracownik</div>
-    							<div className="row"> {/* placeholder */}
-    								<div className="col">
-    									<div className="card-body align-items-center text-center">
-    										<div className="user-avatar user-avatar-xl fileinput-button">
-    											<div className="fileinput-button-label"> Dodaj/zmień zdjęcie </div><img src="/onboarding/static/images/unknown-profile.jpg" alt="" />
-    											<input id="fileupload-avatar" type="file" name="avatar" />
-    										</div>
-    									</div>
-    								</div>
-
-    								<div className="col">
-    									<UserProfileManage user={ singleUser } />
-    								</div>
-    							</div>
+    							<EmployeeEditForm user={ singleUser } />
     						</div>
     					</div>
     				</div>
